# docs/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Luqa Mobile API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Stable native-client contract for Luqa. Authentication uses revocable
    opaque bearer tokens; the browser companion retains its Auth.js session.
servers:
  - url: /api/v1
...
paths:
...
  /gym/locations:
    post:
      operationId: createGymLocation
      tags: [Gym]
      summary: Add a gym
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGymLocationRequest"
      responses:
        "200":
          description: >-
            A gym with this code already exists; it is returned rather than
            duplicated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymLocationResponse"
        "201":
          description: Gym created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymLocationResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /gym/locations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateGymLocation
      tags: [Gym]
      summary: Edit, archive, or restore a gym
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGymLocationRequest"
      responses:
        "200":
          description: Gym updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymLocationResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /money/expenses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateExpense
      tags: [Money]
      summary: Edit a bill
      description: >-
        Omitted fields stay unchanged. Anything that can move the numbers
        re-runs the split; fields left out of a re-split fall back to how the
        bill already looks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseRequest"
      responses:
        "200":
          description: Saved bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteExpense
      tags: [Money]
      summary: Drop a bill and every share on it
      responses:
        "204":
          description: >-
            Gone. Also the answer for a bill that was already deleted, so a
            queued delete replayed from a device cannot fail on its second try.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /money/groups:
    get:
      operationId: listGroups
      tags: [Money]
      summary: List groups with their member ids
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createGroup
      tags: [Money]
      summary: Create a group from a set of people
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "200":
          description: >-
            A group with this name — or this id — already exists; it is
            returned rather than duplicated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/IdConflict"
  /money/groups/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateGroup
      tags: [Money]
      summary: Rename, restyle, change membership, archive, or restore a group
      description: >-
        Sending `memberIds` replaces the membership wholesale; omitting it
        leaves the members alone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGroup
      tags: [Money]
      summary: Remove a group
      description: >-
        Past bills keep their people and amounts; they simply lose the label.
      responses:
        "204":
          description: Gone, or already gone.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    mobileBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived Luqa mobile access token.
  responses:
    InvalidInput:
      description: Malformed JSON, input, or query parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Invalid, expired, revoked, or missing device credential
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The resource does not exist, or is not owned by this user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    IdConflict:
      description: The client-supplied id already belongs to another account
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    GymLocation:
      type: object
      additionalProperties: false
      required: [id, code, name, color, order, archived]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        order:
          type: integer
        archived:
          type: boolean
...
    GymLocationResponse:
      type: object
      additionalProperties: false
      required: [location]
      properties:
        location:
          $ref: "#/components/schemas/GymLocation"
...
    CreateGymLocationRequest:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        id:
          type: string
          minLength: 8
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
          description: >-
            Preferred identity for a gym added offline. Honoured only when free; a gym with the same code already existing wins, so the response is authoritative.
        code:
          type: string
          minLength: 1
          maxLength: 12
        name:
          type: string
          minLength: 1
          maxLength: 60
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
    UpdateGymLocationRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 12
        name:
          type: string
          minLength: 1
          maxLength: 60
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        order:
          type: integer
          minimum: 0
        archived:
          type: boolean
    SplitMode:
      type: string
      description: >-
        How the per-person shares were arrived at. Stored so reopening a bill
        shows the editor state it was saved with.
      enum: [EQUAL, PERCENT, AMOUNT]
...
    PersonGroup:
      type: object
      additionalProperties: false
      required: [id, name, color, emoji, order, archived, memberIds]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        emoji:
          type: [string, "null"]
        order:
          type: integer
        archived:
          type: boolean
        memberIds:
          type: array
          items:
            type: string
    ExpenseShare:
      type: object
      additionalProperties: false
      required: [personId, amountCents, percentBp, gifted]
      properties:
        personId:
          type: string
        amountCents:
          type: integer
        percentBp:
          type: [integer, "null"]
          description: This slice as basis points of the bill; 10000 is 100%.
        gifted:
          type: boolean
          description: >-
            Covered as a treat. Still totalled against the person, never a debt.
    Expense:
      type: object
      additionalProperties: false
      required:
        - id
        - description
        - amountCents
        - date
        - paidByPersonId
        - groupId
        - splitMode
        - myShareCents
        - notes
        - shares
        - createdAt
      properties:
        id:
          type: string
        description:
          type: string
        amountCents:
          type: integer
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        paidByPersonId:
          type: [string, "null"]
          description: >-
            Who fronted the money. Null means the user did. When someone else
            paid, only the user's own share moves a balance.
        groupId:
          type: [string, "null"]
        splitMode:
          $ref: "#/components/schemas/SplitMode"
        myShareCents:
          type: integer
          description: The user's own slice. Shares plus this equals the bill.
        notes:
          type: string
        shares:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseShare"
        createdAt:
          type: string
          format: date-time
...
    ExpenseResponse:
      type: object
      additionalProperties: false
      required: [expense]
      properties:
        expense:
          $ref: "#/components/schemas/Expense"
...
    GroupResponse:
      type: object
      additionalProperties: false
      required: [group]
      properties:
        group:
          $ref: "#/components/schemas/PersonGroup"
    GroupListResponse:
      type: object
      additionalProperties: false
      required: [groups]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/PersonGroup"
...
    ExpenseParticipantInput:
      type: object
      additionalProperties: false
      required: [personId]
      properties:
        personId:
          type: string
        percentBp:
          type: [integer, "null"]
          minimum: 0
          maximum: 10000
          description: "PERCENT mode: this person's cut, in basis points."
        amountCents:
          type: [integer, "null"]
          minimum: 0
          maximum: 100000000
          description: "AMOUNT mode: the exact cents this person carries."
        gifted:
          type: boolean
...
    UpdateExpenseRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        description:
          type: string
          maxLength: 200
        amountCents:
          type: integer
          minimum: 1
          maximum: 100000000
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        paidByPersonId:
          type: [string, "null"]
        groupId:
          type: [string, "null"]
        splitMode:
          $ref: "#/components/schemas/SplitMode"
        includeMe:
          type: boolean
        participants:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/ExpenseParticipantInput"
        notes:
          type: string
          maxLength: 2000
...
    CreateGroupRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id:
          type: string
          minLength: 8
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
          description: >-
            Preferred identity for a group made offline. Honoured only when free; an existing group with the same name wins.
        name:
          type: string
          minLength: 1
          maxLength: 60
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        emoji:
          type: [string, "null"]
          maxLength: 8
        memberIds:
          type: array
          maxItems: 50
          items:
            type: string
    UpdateGroupRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        emoji:
          type: [string, "null"]
          maxLength: 8
        memberIds:
          type: array
          maxItems: 50
          items:
            type: string
        order:
          type: integer
          minimum: 0
        archived:
          type: boolean
...
    ErrorDetail:
      type: object
      additionalProperties: true
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        issues: {}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
security:
  - mobileBearer: []